Remove unused using directives
Removed unnecessary using directives across multiple files in the controllers, services, and repository directories. This is aligned with the general objective of keeping the code clean and minimal. Such measures aid in improving the readability of the code and reduce potential confusion due to unused namespaces. A few key examples are the removal of 'Microsoft.AspNetCore.Http.HttpResults', 'System.Linq.Expressions' and 'Microsoft.EntityFrameworkCore' among others.